requirements.txt
numpy==1.24.1
Pillow==9.4.0
pygame==2.1.2
PyOpenGL==3.1.6
PyQt6==6.4.0
PySide6==6.4.2
PyGLM==2.7.0
